Check raw_body instead of body on initialize
We're setting a default body if there isn't currently anything there.

OutgoingMessage#body contains a load of work to redact the message text,
and gets called every time an OutgoingMessage is initialized. So, even
when you just want to retrieve one and not call #body, it still ends up
doiing all the redaction.

Directly reading the database attribute stops us doing all the redaction
work currently in #body until we need it.